Rearrange some code
